Separate variance baseline service
Baseline creation and defaulting had grown large enough to obscure the remaining variance calculation and alert responsibilities. This keeps server/services/variance-tracking as the public facade while moving baseline-specific imports, helpers, and the shared unique-constraint guard into focused modules.

Constraint: Preserve server/services/variance-tracking imports for routes, automation, and tests

Rejected: Add server/services/variance-tracking/index.ts | same-name file/directory resolver risk

Rejected: Extract alert or calculation services in same commit | review surface would exceed requested baseline slice

Confidence: high

Scope-risk: narrow

Directive: Do not import from the facade inside variance-tracking submodules; depend on sibling helpers directly

Tested: npm test -- tests/unit/services/variance-tracking.test.ts (90 passed)

Tested: npm test -- tests/unit/services/baseline-idempotency.test.ts tests/unit/services/system-actor.test.ts tests/unit/services/metrics-attribution.test.ts tests/unit/services/variance-alert-automation.test.ts (53 passed)

Tested: npm test -- tests/unit/api/variance-tracking-api.test.ts (61 passed)

Tested: npm run lint

Tested: npm run check (0 TypeScript errors)

Tested: npm run build

Not-tested: Full npm test suite
