+++ server/services/variance-tracking/db-error-helpers.ts
@@ -1,0 +1,12 @@
+export function isUniqueConstraintViolation(error: unknown, constraintName: string): boolean {
+  if (!error || typeof error !== 'object') {
+    return false;
+  }
+
+  const candidate = error as { code?: string; constraint?: string; message?: string };
+  return (
+    candidate.code === '23505' &&
+    (candidate.constraint === constraintName ||
+      candidate.message?.includes(constraintName) === true)
+  );
+}
